+++ common/src/com/giancola/battleship/net/DataModel.kt
@@ -36,6 +36,8 @@ data class NotificationShotPerformed(val shooter: PlayerId, val gridX: Int, val 
 @Serializable
 data class NotificationGameFinished(val winner: PlayerId): RemoteNotification()
 
+@Serializable
+class NotificationGameDisconnected(): RemoteNotification()